simple_mdp refactor
